+++ packages/core/src/services/dependencies-mocker.spec.ts
@@ -1,5 +1,8 @@
-import { DependenciesMocker } from './dependencies-mocker';
-import { Type } from '@automock/types';
+import { DependenciesReflector } from '@automock/common';
+import { DependenciesMocker, MockedDependencies } from './dependencies-mocker';
+
+import MockedFn = jest.MockedFn;
+import { StubbedInstance, Type } from '@automock/types';
 
 class SomeClassOne {}
 
@@ -11,7 +14,9 @@ const MOCKED = '__MOCKED__';
 describe('Dependencies Mocker Unit Spec', () => {
   let underTest: DependenciesMocker;
   const mockFunctionStub = () => MOCKED;
-  const reflectorMock = { reflectDependencies: jest.fn() };
+  const reflectorMock = {
+    reflectDependencies: jest.fn() as MockedFn<DependenciesReflector['reflectDependencies']>,
+  };
 
   beforeAll(() => {
     underTest = new DependenciesMocker(reflectorMock, mockFunctionStub);
@@ -19,31 +24,47 @@ describe('Dependencies Mocker Unit Spec', () => {
 
   describe('mocking all class dependencies', () => {
     beforeAll(() => {
-      reflectorMock.reflectDependencies.mockReturnValue(
-        new Map<Type | string, Type>([
+      reflectorMock.reflectDependencies.mockReturnValue({
+        constructor: [
           [SomeClassOne, SomeClassOne],
           [SomeClassTwo, SomeClassTwo],
+          [SomeClassTwo, SomeClassTwo],
           ['TOKEN', ClassFromToken],
-        ])
-      );
+        ],
+      });
     });
 
     describe('assuming there is only one mocked class in the given dependencies', () => {
-      class DoesntMatterClass {}
+      let result: MockedDependencies;
+
+      class ArbitraryClass {}
       const alreadyMockedDependencies = new Map<Type | string, Type>([
         [SomeClassTwo, SomeClassTwoMockedLike],
       ]);
 
-      it('should return a map of class dependencies and mock anything not mocked', () => {
-        const result = underTest.mockAllDependencies(DoesntMatterClass)(alreadyMockedDependencies);
+      beforeAll(() => {
+        result = underTest.mockAllDependencies(ArbitraryClass)(alreadyMockedDependencies);
+      });
 
-        expect(result).toEqual(
-          new Map<Type | string, Type | string>([
+      it('should return a map of the mocks of the class dependencies', () => {
+        expect(result.mocks).toEqual<MockedDependencies['mocks']>(
+          new Map<Type | string, StubbedInstance<unknown>>([
             [SomeClassOne, MOCKED],
             [SomeClassTwo, SomeClassTwoMockedLike],
             ['TOKEN', MOCKED],
           ])
         );
+      });
+
+      it('should also return the original reflected dependencies', () => {
+        expect(result.origin).toEqual<MockedDependencies['origin']>({
+          constructor: [
+            [SomeClassOne, SomeClassOne],
+            [SomeClassTwo, SomeClassTwo],
+            [SomeClassTwo, SomeClassTwo],
+            ['TOKEN', ClassFromToken],
+          ],
+        });
       });
     });
   });