+++ packages/common/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 2.0.0 (2023-07-01)
+
+### Code Refactoring
+
+* **common:** change dependencies reflector signature ([96d33a2](https://github.com/omermorad/automock/commit/96d33a28c97ad93c8bd27d50b4bdf8ab43d11308))
+
+### BREAKING CHANGES
+
+* **common:** The `reflectDependencies` method of `DependenciesReflector` now returns a different interface to accommodate the support for both constructor parameters and properties.
+
 # 1.2.0 (2023-06-10)
 
 ### Features